// src/generated/state.ts
export enum Otp {
  AUTHEN = 'AUTHEN',
  CHANGE_PIN = 'CHANGE_PIN',
  SUMMARY_EMS = 'SUMMARY_EMS',
  SUMMARY_SELF = 'SUMMARY_SELF',
  DETAIL_TRADE = 'DETAIL_TRADE',
  WITHDRAW = 'WITHDRAW',
}

export enum Pin {
  AUTHEN = 'AUTHEN',
  AUTHEN_CONFIRM = 'AUTHEN_CONFIRM',
  CHANGE_PIN = 'CHANGE_PIN',
  CHANGE_PIN_CONFIRM = 'CHANGE_PIN_CONFIRM',
}